refactor: split caseStore into slices without breaking persist
- Modularized src/state/caseStore.ts into src/state/caseStore/ folder
- Created separate slices for each domain:
  - casesSlice.ts: case management (create, select, update, delete, status)
  - fieldsSlice.ts: canonical field values (set, confirm, edit, get)
  - photosSlice.ts: photo evidence management (add, update, delete, confirm category)
  - aiSlice.ts: AI extractions (add, confirm, dismiss, get pending)
  - teamSlice.ts: team member management (set, add, remove)
  - timelineSlice.ts: timeline events (set, add, update)
  - photoReportSlice.ts: photo report config (selection, captions, layout, options)
  - investigationReportSlice.ts: investigation report (blocks, signatures)
  - auditAndProgressSlice.ts: audit events and progress calculations
  - pdfSlice.ts: PDF generation

- Created types.ts with CaseStore interface
- Created helpers.ts with shared utility functions
- Created index.ts that combines all slices with persist middleware
- Maintained exact same persist configuration (name, storage, partialize)
- Kept useCaseStore, useSelectedCase, useFieldValue, useCaseProgress exports
- caseStore.ts now acts as re-export wrapper for backward compatibility
- All demo data preserved
- npm run build passes successfully
